Refactor(regist-approve): 화장실 등록 신청 수정
오탈자 수정
