+++ toilet_registration_service/build.gradle
@@ -31,7 +31,9 @@ ext {
 dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
+	implementation 'org.springframework.boot:spring-boot-starter-validation'
+	implementation 'org.hibernate:hibernate-core:5.6.0.Final'  // Hibernate 추가
+	implementation 'javax.persistence:javax.persistence-api:2.2'  // JPA API 추가
 	compileOnly 'org.projectlombok:lombok'
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
 	runtimeOnly 'com.h2database:h2'
